Use np.memmap for simple elements in binary files
For simple elements in binary files, .ply format uses the same bytes
on-disk as in-memory.  With `numpy.memmap`, we can take advantage of
this - reads are backed by (buffered) disk, and writes store the changes
in memory (the file-on-disk is read-only).

Motivation:  most photogrammetery packages can export to .ply format,
and at landscape scale - yay drones - the files for a single scene can
be several GB each and collectively too large for memory.  On my test
dataset, avoiding the read-write-to-memmap pattern avoids ~40GB of disk
IO per run :)